pyproject.toml
[project]
name = "cs4ct"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.4",
    "langchain-openai>=1.0.2",
    "langgraph>=1.0.2",
    "numpy>=2.3.4",
    "openai>=2.7.1",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "supabase==2.24.0",
]
